// src/models/enums/notification-type.enum.ts
export enum NotificationType {
  POST_REPORTED = "POST_REPORTED",
  POST_ARCHIVED = "POST_ARCHIVED",
  POST_APPROVED = "POST_APPROVED",
  POST_LIKED = "POST_LIKED",
  POST_COMMENTED = "POST_COMMENTED",
  COMMENT_REPORTED = "COMMENT_REPORTED",
  COMMENT_ARCHIVED = "COMMENT_ARCHIVED",
  COMMENT_APPROVED = "COMMENT_APPROVED",
  NEW_FOLLOWER = "NEW_FOLLOWER",
  MENTIONED = "MENTIONED",
  SYSTEM_ANNOUNCEMENT = "SYSTEM_ANNOUNCEMENT",
  ACCOUNT_WARNING = "ACCOUNT_WARNING"
} 